pyproject.toml
[project]
name = "statusnewv1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aerich>=0.9.0",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "tomlkit>=0.13.3",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.34.3",
]

[tool.aerich]
tortoise_orm = "src.core.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
